requirements.txt
django=4.2.13
dajngorestframework== 3.15.1
djangorestframework-simplejwt==5.3.1